Dodawanie i usuwanie produktów. Zapisywanie przepisów do bazy. Wyświetlanie i edycja przepisów.

+++ app/src/main/java/pl/cookbook/database/entities/Recipe.java
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey;
 @Entity
 public class Recipe {
     @PrimaryKey(autoGenerate = true)
-    public int idRecipe;
+    public long idRecipe;
 
     public String name;
